fix: refactor App.tsx
